Portal
-can create a portal in the level
-add a vector position of portal in the class Level
- finding a way to perceive if character is colliding with a portal (maybe using portal's position in the level, which means also need player's position in the level class)

+++ src/business/Game.cpp
@@ -148,7 +148,7 @@ void Game::Render()
             for(int YPosition = 0; YPosition < 10; YPosition++) {
                 for(int XPosition =0; XPosition < 15; XPosition++) {
                     
-                    Rectangle dest{SCREEN_OFFSET+ XPosition * 64, SCREEN_OFFSET + YPosition * 64, 64, 64};
+                    Rectangle dest{(float)SCREEN_OFFSET+ (float)XPosition * 64, (float)SCREEN_OFFSET + (float)YPosition * 64, 64, 64};
                     DrawTexturePro(_grid,source,dest,zero,0,WHITE); 
                 }
             }
